requirements.txt
fastapi
uvicorn
alpaca-py
openai
sqlalchemy
python-dotenv
apscheduler
pydantic
requests
